refactor example and documentation
