+++ docs/data/date-pickers/custom-components/PaperContentComponent.js
@@ -2,7 +2,7 @@ import * as React from 'react';
 import PropTypes from 'prop-types';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
-import { DateRangePicker } from '@mui/x-date-pickers-pro';
+import { DateRangePicker, StaticDateRangePicker } from '@mui/x-date-pickers-pro';
 import {
   TextField,
   Box,
@@ -12,6 +12,8 @@ import {
   ListItem,
   ListItemButton,
   ListItemText,
+  Stack,
+  Button,
 } from '@mui/material';
 import {
   startOfWeek,
@@ -58,41 +60,36 @@ const rangeShortcuts = [
   },
 ];
 
+const buildHandleRangeClick = (setValue) => (range) => {
+  const today = new Date();
+  switch (range) {
+    case RangeShortcut.thisWeek:
+      setValue([startOfWeek(today), endOfWeek(today)]);
+      break;
+    case RangeShortcut.lastWeek:
+      setValue([addWeeks(startOfWeek(today), -1), addWeeks(endOfWeek(today), -1)]);
+      break;
+    case RangeShortcut.last7Days:
+      setValue([addWeeks(today, -1), today]);
+      break;
+    case RangeShortcut.currentMonth:
+      setValue([startOfMonth(today), endOfMonth(today)]);
+      break;
+    case RangeShortcut.nextMonth:
+      setValue([addMonths(startOfMonth(today), 1), addMonths(endOfMonth(today), 1)]);
+
+      break;
+    case RangeShortcut.reset:
+      setValue([null, null]);
+      break;
+    default:
+      break;
+  }
+};
+
 const RangeShortcutsPanel = ({ setValue, children }) => {
   const handleRangeClick = React.useCallback(
-    (range) => {
-      const today = new Date();
-      switch (range) {
-        case RangeShortcut.thisWeek:
-          setValue([startOfWeek(today), endOfWeek(today)]);
-          break;
-        case RangeShortcut.lastWeek:
-          setValue([
-            addWeeks(startOfWeek(today), -1),
-            addWeeks(endOfWeek(today), -1),
-          ]);
-
-          break;
-        case RangeShortcut.last7Days:
-          setValue([addWeeks(today, -1), today]);
-          break;
-        case RangeShortcut.currentMonth:
-          setValue([startOfMonth(today), endOfMonth(today)]);
-          break;
-        case RangeShortcut.nextMonth:
-          setValue([
-            addMonths(startOfMonth(today), 1),
-            addMonths(endOfMonth(today), 1),
-          ]);
-
-          break;
-        case RangeShortcut.reset:
-          setValue([null, null]);
-          break;
-        default:
-          break;
-      }
-    },
+    (range) => buildHandleRangeClick(setValue)(range),
     [setValue],
   );
 
@@ -123,6 +120,41 @@ RangeShortcutsPanel.propTypes = {
   setValue: PropTypes.func.isRequired,
 };
 
+const StaticRangeShortcutsPanel = ({ setValue, children }) => {
+  const handleRangeClick = React.useCallback(
+    (range) => buildHandleRangeClick(setValue)(range),
+    [setValue],
+  );
+
+  return (
+    <React.Fragment>
+      <Box sx={{ m: 2 }}>
+        <Typography variant="overline">Date range shortcuts</Typography>
+        <Box display="flex" gap={2} my={2}>
+          {rangeShortcuts.map(({ range, label }) => (
+            <Button
+              key={range}
+              onClick={() => handleRangeClick(range)}
+              variant="text"
+              color="inherit"
+              size="small"
+            >
+              {label}
+            </Button>
+          ))}
+        </Box>
+        <Divider />
+      </Box>
+      {children}
+    </React.Fragment>
+  );
+};
+
+StaticRangeShortcutsPanel.propTypes = {
+  children: PropTypes.node,
+  setValue: PropTypes.func.isRequired,
+};
+
 export default function PaperContentComponent() {
   const [value, setValue] = React.useState([null, null]);
   const WrappedPaperContent = React.useCallback(
@@ -132,21 +164,39 @@ export default function PaperContentComponent() {
     [],
   );
 
+  const WrappedStaticPaperContent = React.useCallback(
+    ({ children }) => (
+      <StaticRangeShortcutsPanel setValue={setValue}>
+        {children}
+      </StaticRangeShortcutsPanel>
+    ),
+    [],
+  );
+
   return (
     <LocalizationProvider dateAdapter={AdapterDateFns}>
-      <DateRangePicker
-        onChange={(newValue) => setValue(newValue)}
-        value={value}
-        renderInput={(startProps, endProps) => (
-          <React.Fragment>
-            <TextField {...startProps} />
-            <Box sx={{ mx: 2 }}> to </Box>
-            <TextField {...endProps} />
-          </React.Fragment>
-        )}
-        components={{ PaperContent: WrappedPaperContent }}
-        PaperProps={{ sx: { display: 'flex', flexDirection: 'row' } }}
-      />
+      <Stack spacing={4} alignItems="center">
+        <DateRangePicker
+          onChange={(newValue) => setValue(newValue)}
+          value={value}
+          renderInput={(startProps, endProps) => (
+            <React.Fragment>
+              <TextField {...startProps} />
+              <Box sx={{ mx: 2 }}> to </Box>
+              <TextField {...endProps} />
+            </React.Fragment>
+          )}
+          components={{ PaperContent: WrappedPaperContent }}
+          PaperProps={{ sx: { display: 'flex', flexDirection: 'row' } }}
+        />
+        <StaticDateRangePicker
+          displayStaticWrapperAs="desktop"
+          onChange={(newValue) => setValue(newValue)}
+          value={value}
+          renderInput={() => <div />}
+          components={{ PaperContent: WrappedStaticPaperContent }}
+        />
+      </Stack>
     </LocalizationProvider>
   );
 }
