+++ docs/data/date-pickers/custom-components/PaperContentComponent.tsx
@@ -1,7 +1,11 @@
 import * as React from 'react';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
-import { DateRange, DateRangePicker } from '@mui/x-date-pickers-pro';
+import {
+  DateRange,
+  DateRangePicker,
+  StaticDateRangePicker,
+} from '@mui/x-date-pickers-pro';
 import {
   TextField,
   Box,
@@ -11,6 +15,8 @@ import {
   ListItem,
   ListItemButton,
   ListItemText,
+  Stack,
+  Button,
 } from '@mui/material';
 import {
   startOfWeek,
@@ -59,42 +65,43 @@ const rangeShortcuts = [
   },
 ];
 
+const buildHandleRangeClick =
+  (setValue: React.Dispatch<React.SetStateAction<DateRange<Date>>>) =>
+  (range: RangeShortcutType) => {
+    const today = new Date();
+    switch (range) {
+      case RangeShortcut.thisWeek:
+        setValue([startOfWeek(today), endOfWeek(today)]);
+        break;
+      case RangeShortcut.lastWeek:
+        setValue([addWeeks(startOfWeek(today), -1), addWeeks(endOfWeek(today), -1)]);
+        break;
+      case RangeShortcut.last7Days:
+        setValue([addWeeks(today, -1), today]);
+        break;
+      case RangeShortcut.currentMonth:
+        setValue([startOfMonth(today), endOfMonth(today)]);
+        break;
+      case RangeShortcut.nextMonth:
+        setValue([
+          addMonths(startOfMonth(today), 1),
+          addMonths(endOfMonth(today), 1),
+        ]);
+        break;
+      case RangeShortcut.reset:
+        setValue([null, null]);
+        break;
+      default:
+        break;
+    }
+  };
+
 const RangeShortcutsPanel: React.FC<{
   setValue: React.Dispatch<React.SetStateAction<DateRange<Date>>>;
   children: React.ReactNode;
 }> = ({ setValue, children }) => {
   const handleRangeClick = React.useCallback(
-    (range: RangeShortcutType) => {
-      const today = new Date();
-      switch (range) {
-        case RangeShortcut.thisWeek:
-          setValue([startOfWeek(today), endOfWeek(today)]);
-          break;
-        case RangeShortcut.lastWeek:
-          setValue([
-            addWeeks(startOfWeek(today), -1),
-            addWeeks(endOfWeek(today), -1),
-          ]);
-          break;
-        case RangeShortcut.last7Days:
-          setValue([addWeeks(today, -1), today]);
-          break;
-        case RangeShortcut.currentMonth:
-          setValue([startOfMonth(today), endOfMonth(today)]);
-          break;
-        case RangeShortcut.nextMonth:
-          setValue([
-            addMonths(startOfMonth(today), 1),
-            addMonths(endOfMonth(today), 1),
-          ]);
-          break;
-        case RangeShortcut.reset:
-          setValue([null, null]);
-          break;
-        default:
-          break;
-      }
-    },
+    (range: RangeShortcutType) => buildHandleRangeClick(setValue)(range),
     [setValue],
   );
   return (
@@ -121,6 +128,38 @@ const RangeShortcutsPanel: React.FC<{
   );
 };
 
+const StaticRangeShortcutsPanel: React.FC<{
+  setValue: React.Dispatch<React.SetStateAction<DateRange<Date>>>;
+  children: React.ReactNode;
+}> = ({ setValue, children }) => {
+  const handleRangeClick = React.useCallback(
+    (range: RangeShortcutType) => buildHandleRangeClick(setValue)(range),
+    [setValue],
+  );
+  return (
+    <React.Fragment>
+      <Box sx={{ m: 2 }}>
+        <Typography variant="overline">Date range shortcuts</Typography>
+        <Box display="flex" gap={2} my={2}>
+          {rangeShortcuts.map(({ range, label }) => (
+            <Button
+              key={range}
+              onClick={() => handleRangeClick(range as RangeShortcutType)}
+              variant="text"
+              color="inherit"
+              size="small"
+            >
+              {label}
+            </Button>
+          ))}
+        </Box>
+        <Divider />
+      </Box>
+      {children}
+    </React.Fragment>
+  );
+};
+
 export default function PaperContentComponent() {
   const [value, setValue] = React.useState<DateRange<Date>>([null, null]);
   const WrappedPaperContent = React.useCallback(
@@ -129,21 +168,38 @@ export default function PaperContentComponent() {
     ),
     [],
   );
+  const WrappedStaticPaperContent = React.useCallback(
+    ({ children }: { children: React.ReactNode }) => (
+      <StaticRangeShortcutsPanel setValue={setValue}>
+        {children}
+      </StaticRangeShortcutsPanel>
+    ),
+    [],
+  );
   return (
     <LocalizationProvider dateAdapter={AdapterDateFns}>
-      <DateRangePicker
-        onChange={(newValue) => setValue(newValue)}
-        value={value}
-        renderInput={(startProps, endProps) => (
-          <React.Fragment>
-            <TextField {...startProps} />
-            <Box sx={{ mx: 2 }}> to </Box>
-            <TextField {...endProps} />
-          </React.Fragment>
-        )}
-        components={{ PaperContent: WrappedPaperContent }}
-        PaperProps={{ sx: { display: 'flex', flexDirection: 'row' } }}
-      />
+      <Stack spacing={4} alignItems="center">
+        <DateRangePicker
+          onChange={(newValue) => setValue(newValue)}
+          value={value}
+          renderInput={(startProps, endProps) => (
+            <React.Fragment>
+              <TextField {...startProps} />
+              <Box sx={{ mx: 2 }}> to </Box>
+              <TextField {...endProps} />
+            </React.Fragment>
+          )}
+          components={{ PaperContent: WrappedPaperContent }}
+          PaperProps={{ sx: { display: 'flex', flexDirection: 'row' } }}
+        />
+        <StaticDateRangePicker
+          displayStaticWrapperAs="desktop"
+          onChange={(newValue) => setValue(newValue)}
+          value={value}
+          renderInput={() => <div />}
+          components={{ PaperContent: WrappedStaticPaperContent }}
+        />
+      </Stack>
     </LocalizationProvider>
   );
 }
