--- docs/data/date-pickers/custom-components/PaperContentComponent.tsx.preview
@@ -1,15 +1,0 @@
-<LocalizationProvider dateAdapter={AdapterDateFns}>
-  <DateRangePicker
-    onChange={(newValue) => setValue(newValue)}
-    value={value}
-    renderInput={(startProps, endProps) => (
-      <React.Fragment>
-        <TextField {...startProps} />
-        <Box sx={{ mx: 2 }}> to </Box>
-        <TextField {...endProps} />
-      </React.Fragment>
-    )}
-    components={{ PaperContent: WrappedPaperContent }}
-    PaperProps={{ sx: { display: 'flex', flexDirection: 'row' } }}
-  />
-</LocalizationProvider>
