+++ docs/data/date-pickers/custom-components/custom-components.md
@@ -83,11 +83,10 @@ In the example below, the actions are the same as in the section above, but they
 
 ## Paper content
 
-The paper content is available on all desktop picker components.  
-It adds a flexible way to extend what is rendered in the picker paper.  
-You can provide any custom component to this slot as long as it passes `children` down maintaining original picker behavior.  
+The paper content is available on all desktop picker components.
+It adds a flexible way to extend what is rendered in the picker paper.
+
+You can provide any custom component to this slot as long as it passes `children` down maintaining original picker behavior.
 In the examples below we provide ways of implementing date range shortcuts using this slot.
 
-{{"demo": "PaperContentComponent.js"}}
-
-{{"demo": "StaticPaperContentComponent.js"}}
+{{"demo": "PaperContentComponent.js", "defaultCodeOpen": false, "disableAd": true}}
